requirements.txt
python-telegram-bot==20.7
openai
feedparser
pandas
requests
beautifulsoup4
schedule
python-dotenv